add new tests

+++ mojo_curl/http.mojo
@@ -1,10 +1,10 @@
-from std.ffi import c_int
+from std.ffi import c_long
 
 @fieldwise_init
 struct HTTPVersion(TrivialRegisterPassable):
     """CURL HTTP version options for setting HTTP versions."""
 
-    var value: c_int
+    var value: c_long
     comptime NONE = Self(0)
     comptime V1_0 = Self(1)
     comptime V1_1 = Self(2)
